# src/node-server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.1",
    "title": "José Neto Thesis API",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module.\n                    <p> José Neto </p>"
  },
  "host": "localhost:5431",
  "basePath": "/",
  "tags": [
    {
      "name": "Measurements",
      "description": "Endpoints to upload, read or delete measurement data"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/measurements/multiple": {
      "post": {
        "tags": [
          "Measurements"
        ],
        "description": "Endpoint to upload sensor data.",
        "parameters": [
          {
            "name": "sensorData",
            "in": "body",
            "description": "Measurement from installation",
            "required": true,
            "type": "object"
          }
        ],
        "responses": {}
      }
    },
    "/measurements": {
      "get": {
        "tags": [
          "Measurements"
        ],
        "description": "Returns all measurements.",
        "parameters": [],
        "responses": {}
      }
    },
    "/measurements/{sensor}": {
      "get": {
        "tags": [
          "Measurements"
        ],
        "description": "Gets data from specific sensor.",
        "parameters": [
          {
            "name": "sensor",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sensorData",
            "in": "body",
            "description": "Measurement from installation",
            "required": true,
            "type": "object"
          }
        ],
        "responses": {}
      }
    },
    "/measurements/today": {
      "get": {
        "tags": [
          "Measurements"
        ],
        "description": "Gets data from today.",
        "parameters": [
          {
            "name": "sensorData",
            "in": "body",
            "description": "Measurement from today",
            "required": true,
            "type": "object"
          }
        ],
        "responses": {}
      }
    },
    "/measurements/{id}": {
      "get": {
        "tags": [
          "Measurements"
        ],
        "description": "Gets specific measurement.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sensorData",
            "in": "body",
            "description": "Gets specific measurement",
            "required": true,
            "type": "object"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Measurements"
        ],
        "description": "Deletes specific measurement.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sensorData",
            "in": "body",
            "description": "Deletes specific measurement",
            "required": true,
            "type": "object"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {}
}